Iterate on deconvolution notebooks
The main change is that the 10_decon_systems_design notebook has been
tweaked to show both the subotpimal and optimal deconvolution.

This makes it easier to show the difference between deconvolutions done
with different parameters.  However if using the notebook for a course,
it would be a good idea to hide the optimal deconvolution result, and
let the students try to figure it out as a excercise.